Handle 1.13.2->1.14 lore tag properly
Handle the lore tag like 1.14 does by checking for null components / removing the entire lore tag when one of its entries is invalid; I believe we would also need this for name conversion one protocol below. The Problem is that we need to handle translations there, and since the current component rewriter base relies on JsonElement, we would need to parse the name multiple times, so I left this out.