Fix text components in signs not being rewritten in the 1.21.9→1.21.7 step
This was causing connection crashes for versions ≤1.21.4, since the
1.21.5.→1.21.4 step rewrites all text components as strings, and thus
parses the NBT.
Since sprites and other object components were not being filtered on
signs in the 1.21.9→1.21.7 step (despite being correctly handled
everywhere else since #1108), the latter step was failing to properly
read the NBT structure and thus crashing the player's connection.

The fix was done by lifting the sign logic in
`BlockItemPacketRewriter1_21_5` and adapting it into
`BlockItemPacketRewriter1_21_9`. All filtering is still delegated to the
appropriate `ComponentRewriter`, so required code change was rather
small.